+++ mobile/android/android-components/components/concept/storage/src/main/java/mozilla/components/concept/storage/CreditCardsAddressesStorage.kt
@@ -6,8 +6,11 @@ package mozilla.components.concept.storage
 
 import android.annotation.SuppressLint
 import android.os.Parcelable
-import kotlinx.coroutines.Deferred
 import kotlinx.parcelize.Parcelize
+import mozilla.components.concept.storage.CreditCard.Companion.ellipsesEnd
+import mozilla.components.concept.storage.CreditCard.Companion.ellipsesStart
+import mozilla.components.concept.storage.CreditCard.Companion.ellipsis
+import mozilla.components.support.ktx.kotlin.last4Digits
 
 /**
  * An interface which defines read/write methods for credit card and address data.
@@ -49,6 +52,7 @@ interface CreditCardsAddressesStorage {
     /**
      * Deletes the credit card with the given [guid].
      *
+     * @param guid Unique identifier for the desired credit card.
      * @return True if the deletion did anything, false otherwise.
      */
     suspend fun deleteCreditCard(guid: String): Boolean
@@ -196,10 +200,10 @@ data class CreditCard(
     val expiryMonth: Long,
     val expiryYear: Long,
     val cardType: String,
-    val timeCreated: Long,
-    val timeLastUsed: Long?,
-    val timeLastModified: Long,
-    val timesUsed: Long
+    val timeCreated: Long = 0L,
+    val timeLastUsed: Long? = 0L,
+    val timeLastModified: Long = 0L,
+    val timesUsed: Long = 0L
 ) : Parcelable {
     val obfuscatedCardNumber: String
         get() = ellipsesStart +
@@ -217,6 +221,33 @@ data class CreditCard(
         // Pop Directional Formatting (PDF) mark
         const val ellipsesEnd = "\u202C"
     }
+}
+
+/**
+ * Credit card autofill entry.
+ *
+ * This contains the data needed to handle autofill but not the data related to the DB record.
+ *
+ * @property guid The unique identifier for this credit card.
+ * @property name The credit card billing name.
+ * @property number The credit card number.
+ * @property expiryMonth The credit card expiry month.
+ * @property expiryYear The credit card expiry year.
+ * @property cardType The credit card network ID.
+ */
+data class CreditCardEntry(
+    val guid: String? = null,
+    val name: String,
+    val number: String,
+    val expiryMonth: String,
+    val expiryYear: String,
+    val cardType: String
+) {
+    val obfuscatedCardNumber: String
+        get() = ellipsesStart +
+            ellipsis + ellipsis + ellipsis + ellipsis +
+            number.last4Digits() +
+            ellipsesEnd
 }
 
 /**
@@ -335,39 +366,84 @@ data class UpdatableAddressFields(
 )
 
 /**
+ * Provides a method for checking whether or not a given credit card can be stored.
+ */
+interface CreditCardValidationDelegate {
+
+    /**
+     * The result from validating a given [CreditCard] against the credit card storage. This will
+     * include whether or not it can be created or updated.
+     */
+    sealed class Result {
+        /**
+         * Indicates that the [CreditCard] does not currently exist in the storage, and a new
+         * credit card entry can be created.
+         */
+        object CanBeCreated : Result()
+
+        /**
+         * Indicates that a matching [CreditCard] was found in the storage, and the [CreditCard]
+         * can be used to update its information.
+         */
+        data class CanBeUpdated(val foundCreditCard: CreditCard) : Result()
+    }
+
+    /**
+     * Determines whether a [CreditCardEntry] can be added or updated in the credit card storage.
+     *
+     * @param creditCard [CreditCardEntry] to be added or updated in the credit card storage.
+     * @return [Result] that indicates whether or not the [CreditCardEntry] should be saved or
+     * updated.
+     */
+    suspend fun shouldCreateOrUpdate(creditCard: CreditCardEntry): Result
+}
+
+/**
  * Used to handle [Address] and [CreditCard] storage so that the underlying engine doesn't have to.
  * An instance of this should be attached to the Gecko runtime in order to be used.
  */
-interface CreditCardsAddressesStorageDelegate {
+interface CreditCardsAddressesStorageDelegate : KeyProvider {
 
     /**
      * Decrypt a [CreditCardNumber.Encrypted] into its plaintext equivalent or `null` if
      * it fails to decrypt.
      *
+     * @param key The encryption key to decrypt the decrypt credit card number.
      * @param encryptedCardNumber An encrypted credit card number to be decrypted.
      * @return A plaintext, non-encrypted credit card number.
      */
-    suspend fun decrypt(encryptedCardNumber: CreditCardNumber.Encrypted): CreditCardNumber.Plaintext?
+    suspend fun decrypt(
+        key: ManagedKey,
+        encryptedCardNumber: CreditCardNumber.Encrypted
+    ): CreditCardNumber.Plaintext?
 
     /**
      * Returns all stored addresses. This is called when the engine believes an address field
      * should be autofilled.
+     *
+     * @return A list of all stored addresses.
      */
-    fun onAddressesFetch(): Deferred<List<Address>>
+    suspend fun onAddressesFetch(): List<Address>
 
     /**
      * Saves the given address to storage.
+     *
+     * @param address [Address] to be saved or updated in the address storage.
      */
-    fun onAddressSave(address: Address)
+    suspend fun onAddressSave(address: Address)
 
     /**
      * Returns all stored credit cards. This is called when the engine believes a credit card
      * field should be autofilled.
+     *
+     * @return A list of all stored credit cards.
      */
-    fun onCreditCardsFetch(): Deferred<List<CreditCard>>
+    suspend fun onCreditCardsFetch(): List<CreditCard>
 
     /**
      * Saves the given credit card to storage.
+     *
+     * @param creditCard [CreditCardEntry] to be saved or updated in the credit card storage.
      */
-    fun onCreditCardSave(creditCard: CreditCard)
+    suspend fun onCreditCardSave(creditCard: CreditCardEntry)
 }